+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 - PNDA-2456: Initial work to support for Redhat 7. Salt highstate and orchestrate run on a Redhat7 HEAT cluster with no errors but requires further testing and work.
 - PNDA-2801: Add support for bare-metal deployment using the bmstandard flavor. Add support for distribution flavor providing kafka only cluser. Add documentation for baremetal deployment.
+- Add an hypervisor_count setting in the pnda_env file to enable Anti-Affinity feature in PNDA.
 ### Changed
 - PNDA-2688: review pnda_env default values
 - PNDA-2819: fix issue on volume reference once create network is 0